Errors Fix

+++ integer_operation.c
@@ -23,8 +23,8 @@ int op_integer(va_list argz)
 
 int d_i_print(va_list argz)
 {
-	int digit, count, temp;
-	int place_holder, whole;
+	long int digit, count, temp;
+	long int place_holder, whole;
 
 	place_holder = 1;
 	whole = count = temp = digit = 0;
